refactor: move select components to component dir
